crm_automation_bot/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
pandas==2.1.4
beautifulsoup4==4.12.2
python-suitecrm==0.1.0
schedule==1.2.1 